+++ backend/src/cooking/Cook.cpp
@@ -1,15 +1,16 @@
 #include "Cook.h"
 #include "Baker.h"
-#include "Order.h"
 #include "Meal.h"
-#include "MenuItem.h"
+
+#include "../ordering/Order.h"
+#include "../ordering/MenuItem.h"
 
 Cook::Cook(Kitchen* Kitchen) : StationChef("Cook",kitchen){}
 
 Cook::~Cook(){}
 
 void Cook::prepareOrder(std::shared_ptr<Order> order){
-    for(int i = 0; i < order->getMeals().size(); i++){
+    for(int i = 0; i < (int) order->getMeals().size(); i++){
         if(this->kitchen->getChefName(order->getMeals()[i]->getName())==this->getName()){
             order->getMeals()[i]->prepare();
         }
