+++ backend/src/cooking/GrillChef.cpp
@@ -1,8 +1,9 @@
 #include "GrillChef.h"
 #include "Cook.h"
-#include "Order.h"
 #include "Meal.h"
-#include "MenuItem.h"
+
+#include "../ordering/Order.h"
+#include "../ordering/MenuItem.h"
 
 GrillChef::GrillChef(Kitchen* kitchen):StationChef("Grill Chef",kitchen){}
 
@@ -10,7 +11,7 @@ GrillChef::~GrillChef(){}
 
 void GrillChef::prepareOrder(std::shared_ptr<Order> order){
     if(order!=nullptr){
-        for(int i = 0; i < order->getMeals().size(); i++){
+        for(int i = 0; i < (int) order->getMeals().size(); i++){
             if(this->kitchen->getChefName(order->getMeals()[i]->getName())==this->getName()){
                 order->getMeals()[i]->prepare();
             }
